+++ src/overlays/ray.ts
@@ -8,26 +8,23 @@ const ray: OverlayTemplate = {
   needDefaultYAxisFigure: true,
   createPointFigures: ({ coordinates, bounding }: any) => {
     if (coordinates.length > 1) {
-      let coordinate = { x: 0, y: 0 };
-      if (
-        coordinates[0].x === coordinates[1].x &&
-        coordinates[0].y !== coordinates[1].y
-      ) {
-        if (coordinates[0].y < coordinates[1].y) {
-          coordinate = { x: coordinates[0].x, y: bounding.height };
-        } else {
-          coordinate = { x: coordinates[0].x, y: 0 };
+      const coordinate = (() => {
+        if (
+          coordinates[0].x === coordinates[1].x &&
+          coordinates[0].y !== coordinates[1].y
+        ) {
+          return coordinates[0].y < coordinates[1].y
+            ? { x: coordinates[0].x, y: bounding.height }
+            : { x: coordinates[0].x, y: 0 };
         }
-      } else {
-        const x =
-          coordinates[0].x < coordinates[1].x ? bounding.width : 0;
+        const x = coordinates[0].x < coordinates[1].x ? bounding.width : 0;
         const y =
           ((coordinates[1].y - coordinates[0].y) /
             (coordinates[1].x - coordinates[0].x)) *
             (x - coordinates[0].x) +
           coordinates[0].y;
-        coordinate = { x, y };
-      }
+        return { x, y };
+      })();
       return [
         {
           type: "line",